feat: :truck: Move activity builder and refactor CategoryCosplayDAO
