--- src/main/java/com/u2team/huansync/activitiy/controller/ActivityController.java
@@ -1,5 +1,0 @@
-package com.u2team.huansync.activitiy.controller;
-
-public class ActivityController {
-    
-}
