--- src/main/java/com/u2team/huansync/activitiy/cosplay/controller/CosplayController.java
@@ -1,5 +1,0 @@
-package com.u2team.huansync.activitiy.cosplay.controller;
-
-public class CosplayController {
-    
-}
