--- src/main/java/com/u2team/huansync/activitiy/cosplay/model/Cosplay.java
@@ -1,5 +1,0 @@
-package com.u2team.huansync.activitiy.cosplay.model;
-
-public class Cosplay {
-    
-}
