--- src/main/java/com/u2team/huansync/activitiy/trivia/controller/TriviaController.java
@@ -1,5 +1,0 @@
-package com.u2team.huansync.activitiy.trivia.controller;
-
-public class TriviaController {
-    
-}
