--- src/main/java/com/u2team/huansync/activitiy/trivia/model/Trivia.java
@@ -1,5 +1,0 @@
-package com.u2team.huansync.activitiy.trivia.model;
-
-public class Trivia {
-    
-}
